publish-to-github: Use 'test' instead of 'ls' to check build/
And unwind the one-liner into less-compact if statement.  This makes
the handling more explicit (e.g. it also errors out if there's a build
file).  I also dropped the && between the echo and the exit.  Even if
we hadn't set -e, we'd still want to exit regardless of whether or not
the echo succeeds.