Add /help
This adds an specific page dedicated to funnel users to the right places to
get help. It is meant so that we can link this from everywhere and we have
to just keep a single source of truth updated.

It should reflect current support-ecosystem and it mostly aims to direct users
to the Forums.